Axes are now fixed. Float number can now be entered on the test form.

+++ app/testsimform.py
@@ -1,7 +1,7 @@
 from flask_wtf import Form
-from wtforms import IntegerField
+from wtforms import FloatField
 from wtforms.validators import Required
 
 class TestSimForm(Form):
-    aField = IntegerField('aField', validators = [Required()])
-    mField = IntegerField('mField', validators = [Required()])
+    aField = FloatField('aField', validators = [Required()])
+    mField = FloatField('mField', validators = [Required()])
